Bug: Wave Not Found Averaging; Feature: Rm and Ra
Acquire_main: changed a bug that was resulting in a wave not found error
message. Basically, concat_0 was being called during the averaging, but
was not being made by the do_average() function.

TAB_Utilities: Fixed a minor bug in the oops() function, added a
function to measure the resistance (Ra and Rm) of a 5 mV test pulse